Moved classes related to graphs into engine.graph package
