+++ src/engine/StepperController.java
@@ -1,5 +1,8 @@
 package engine;
 
+import engine.graph.Graph;
+import engine.graph.IGraphRestore;
+
 import java.util.Stack;
 
 public class StepperController
