+++ src/engine/graph/DirectedEdgePair.java
@@ -1,14 +1,14 @@
-package engine;
+package engine.graph;
 
 // directed edges but we don't care what order they are in in the pair
-class DirectedEdgePair
+public class DirectedEdgePair
 {
    final DirectedEdge m_e1;
    final DirectedEdge m_e2;
    final double m_t1;
    final double m_t2;
 
-   DirectedEdgePair(DirectedEdge e1, DirectedEdge e2, double t1, double t2)
+   public DirectedEdgePair(DirectedEdge e1, DirectedEdge e2, double t1, double t2)
    {
       m_e1 = e1;
       m_e2 = e2;
