+++ src/engine/graph/TemplateBuilder.java
@@ -1,4 +1,4 @@
-package engine;
+package engine.graph;
 
 import java.util.*;
