+++ src/engine/graph/TemplateStore.java
@@ -1,4 +1,4 @@
-package engine;
+package engine.graph;
 
 import java.util.ArrayList;
 import java.util.Collection;
