+++ test/engine/UtilTest.java
@@ -98,91 +98,6 @@ public class UtilTest
    }
 
    @Test
-   public void testEdgeIntersectSimple_Node() throws Exception
-   {
-      assertNotNull(Util.edgeIntersect(makeNodeAt(1, 0), makeNodeAt(-1, 0), makeNodeAt(0, 1), makeNodeAt(0, -1)));
-      assertNull(Util.edgeIntersect(makeNodeAt(1, 0), makeNodeAt(-1, 0), makeNodeAt(1, 0), makeNodeAt(-1, 0)));
-
-      assertNotNull(Util.edgeIntersect(makeNodeAt(-1, 0), makeNodeAt(1, 0), makeNodeAt(0, -1), makeNodeAt(0, 1)));
-      assertNull(Util.edgeIntersect(makeNodeAt(-1, 0), makeNodeAt(1, 0), makeNodeAt(-1, 0), makeNodeAt(1, 0)));
-
-      assertNotNull(Util.edgeIntersect(makeNodeAt(0, 1), makeNodeAt(0, -1), makeNodeAt(1, 0), makeNodeAt(-1, 0)));
-      assertNull(Util.edgeIntersect(makeNodeAt(0, 1), makeNodeAt(0, -1), makeNodeAt(0, 1), makeNodeAt(0, -1)));
-
-      assertNotNull(Util.edgeIntersect(makeNodeAt(0, -1), makeNodeAt(0, 1), makeNodeAt(-1, 0), makeNodeAt(1, 0)));
-      assertNull(Util.edgeIntersect(makeNodeAt(0, -1), makeNodeAt(0, 1), makeNodeAt(0, -1), makeNodeAt(0, 1)));
-   }
-
-   @Test
-   public void testEdgeIntersectAdjoining_Node() throws Exception
-   {
-      // we can detect end-collisions of edges
-      assertNotNull(Util.edgeIntersect(new XY(1, 0), new XY(-1, 0), new XY(-1, 0), new XY(0, 1)));
-      assertNotNull(Util.edgeIntersect(new XY(-1, 0), new XY(1, 0), new XY(-1, 0), new XY(0, 1)));
-      assertNotNull(Util.edgeIntersect(new XY(1, 0), new XY(-1, 0), new XY(0, 1), new XY(-1, 0)));
-      assertNotNull(Util.edgeIntersect(new XY(-1, 0), new XY(1, 0), new XY(0, 1), new XY(-1, 0)));
-
-      Node n1 = makeNodeAt(1, 0);
-      Node n2 = makeNodeAt(-1, 0);
-      Node n3 = makeNodeAt(0, 1);
-
-      // but we don't if we know it is deliberate edge concatenation
-      // e.g. if the node is shared
-      assertNull(Util.edgeIntersect(n1, n2, n2, n3));
-      assertNull(Util.edgeIntersect(n2, n1, n2, n3));
-      assertNull(Util.edgeIntersect(n1, n2, n3, n2));
-      assertNull(Util.edgeIntersect(n2, n1, n3, n2));
-   }
-
-   @Test
-   public void testEdgeIntersect_Edges()
-   {
-      {
-         Node n1 = makeNodeAt(1, 0);
-         Node n2 = makeNodeAt(-1, 0);
-         Node n3 = makeNodeAt(0, 1);
-
-         // check same adjoining-edge behaviour as we checked with nodes
-         assertNull(Util.edgeIntersect(makeEdge(n1, n2), makeEdge(n2, n3)));
-         assertNull(Util.edgeIntersect(makeEdge(n2, n1), makeEdge(n2, n3)));
-         assertNull(Util.edgeIntersect(makeEdge(n1, n2), makeEdge(n3, n2)));
-         assertNull(Util.edgeIntersect(makeEdge(n2, n1), makeEdge(n3, n2)));
-      }
-
-      // just repeat a couple of the above tests and check we get the same t values
-      double[] values = {1e-6, 2e-6, 5e-6, 1e-5, 2e-5, 5e-5, 1e-4, 2e-4, 5e-4, 1e-3, 2e-3, 5e-3, 1e-2, 2e-2, 5e-2, 1e-1, 2e-1, 5e-1};
-
-      Node n1 = makeNodeAt(0, 0);
-      Node n2 = makeNodeAt(1, 0);
-      DirectedEdge e1 = makeEdge(n1, n2);
-
-      for (double f : values)
-      {
-         Node n3 = makeNodeAt(f, 0.5);
-         Node n4 = makeNodeAt(f, -0.5);
-         DirectedEdge e2 = makeEdge(n3, n4);
-
-         {
-            DirectedEdgePair ret = Util.edgeIntersect(e1, e2);
-
-            assertEquals(e1, ret.m_e1);
-            assertEquals(e2, ret.m_e2);
-            assertEquals(f, ret.m_t1, 1e-8);
-            assertEquals(0.5, ret.m_t2, 1e-8);
-         }
-
-         {
-            DirectedEdgePair ret = Util.edgeIntersect(e2, e1);
-
-            assertEquals(e2, ret.m_e1);
-            assertEquals(e1, ret.m_e2);
-            assertEquals(0.5, ret.m_t1, 1e-8);
-            assertEquals(f, ret.m_t2, 1e-8);
-         }
-      }
-   }
-
-   @Test
    public void testUnitEdgeForce()
    {
       // between min and max should be zero and no distortion
@@ -458,26 +373,5 @@ public class UtilTest
       assertEquals(Math.PI / 2, Util.relativeAngle(any, any.rot90()), 1e-6);
       assertEquals(Math.PI, Util.relativeAngle(any, any.rot90().rot90()), 1e-6);
       assertEquals(Math.PI *3 / 2, Util.relativeAngle(any, any.rot270()), 1e-6);
-   }
-
-   // ------------------
-
-   private Node makeNodeAt(double x, double y)
-   {
-      return makeRadiusNodeAt(x, y, 0.0);
-   }
-
-   private Node makeRadiusNodeAt(double x, double y, @SuppressWarnings("SameParameterValue") double radius)
-   {
-      Node ret = new Node("", "", "", radius);
-
-      ret.setPos(new XY(x, y));
-
-      return ret;
-   }
-
-   private DirectedEdge makeEdge(Node n1, Node n2)
-   {
-      return new DirectedEdge(n1, n2, 0, 0, 0);
    }
 }
