+++ test/engine/graph/TemplateStoreTest.java
@@ -1,8 +1,5 @@
-package engine;
+package engine.graph;
 
-import engine.Template;
-import engine.TemplateBuilder;
-import engine.TemplateStore;
 import org.junit.Test;
 
 import java.util.Collection;
